ログイン後に画面遷移できるようにした。 そのためにMainActivityに書かれていたコードをLogInFragmentに移動させた。
